4843 Centralize common geoid deserialization
